+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## HEAD
+
+* Add `authenticated?` helper method.
+
+
 ## 1.1.0 (7 October 2011)
 
 * Correctly handle blank username in password reset.